Include spec for index request with pagination params
